Synchronise all contracts-controller calls with vent
Signed-off-by: Silas Davis <silas@monax.io>

Add agreement version 1.0.2 to implement redaction

Add a `REDACTED` legal state available to agreement owners only that extends the cancellation flow. An owner may call `redact()` which will result in setting the legal state to `REDACTED` and emitting a Vent deletion event for the agreement only if:

- `msg.sender` is `owner`

and one of:

1. There are no parties (any legal state)
2. All parties have cancelled and `LegalState < FUFILLED`
2. Any party has cancelled and `LegalState = DRAFT`

It modifies the `CANCELLED` state machine so that in case 1 cancellation
will succeed.

Also adds vent synchronisation to all contract-controller functions

Signed-off-by: Silas Davis <silas@monax.io>